pyproject.toml
[project]
name = "fitness-training"
version = "0.1.0"
description = ""
authors = [
    {name = "Harry Vargas",email = "hvargasr@unal.edu.co"}
]
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.33.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "plotly (>=6.6.0,<7.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "streamlit (>=1.56.0,<2.0.0)",
    "flask (>=3.1.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
